support-flow-nexus-main/backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    # Database
    POSTGRES_USER = os.getenv("POSTGRES_USER", "postgres")
    POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD", "postgres")
    POSTGRES_HOST = os.getenv("POSTGRES_HOST", "localhost")
    POSTGRES_PORT = os.getenv("POSTGRES_PORT", "5432")
    POSTGRES_DB = os.getenv("POSTGRES_DB", "support_flow_nexus")
    
    SQLALCHEMY_DATABASE_URI = f"postgresql://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_DB}"
    
    # JWT
    SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key-for-jwt")
    ALGORITHM = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES = 30
    
    # File Upload
    UPLOAD_FOLDER = "uploads"
    MAX_CONTENT_LENGTH = 16 * 1024 * 1024  # 16MB
    ALLOWED_EXTENSIONS = {"txt", "pdf", "png", "jpg", "jpeg", "gif"}
    
    # CORS
    CORS_ORIGINS = [
        "http://localhost:8080",
        "http://127.0.0.1:8080",
        # Add production URLs when deploying
    ]
    
    # API
    API_V1_STR = "/api"
    PROJECT_NAME = "Support Flow Nexus" 