support-flow-nexus-main/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
python-dotenv
Flask-CORS
bcrypt
python-dateutil
SQLAlchemy
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
email-validator